pyproject.toml
[project]
name = "baq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "chogerlate", email = "siwaratlaopromger4444@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.10.5",
    "dvc>=3.59.1",
    "dvc-s3>=3.2.0",
    "feature-engine>=1.8.3",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "mlflow>=2.21.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "wandb>=0.19.11",
    "xgboost>=3.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/baq"]

[dependency-groups]
lint = [
    "ruff>=0.11.4",
]
dev = [
    "pytest>=8.3.5",
]

